{{ cookiecutter.project_name }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_name }}"
version = "0.1.0"
description = "{{ cookiecutter.project_description }}"
authors = ["{{ cookiecutter.author_name }} <{{ cookiecutter.author_email }}>"]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.11.0"
graphene-django = "^2.9.0"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.4"
whitenoise = "^5.0.1"
model_bakery = "^1.1.0"
drf-yasg = "^1.17.1"
dynaconf = "^2.2.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = "^19.10b0"
isort = "^4.3.21"
bandit = "^1.6.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
